# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Limited.Ink API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface VerifyLicenseBody {
  /** The license activation code */
  code: string;
  /** A hash uniquely identifying the user's device/browser */
  deviceFingerprint: string;
}

export interface CheckLicenseStatusBody {
  /** The JWT session token to validate */
  token: string;
  /** The device fingerprint to match against the token */
  deviceFingerprint: string;
}

export interface LicenseVerifyResponse {
  /** Signed JWT session token */
  token: string;
  /** The plan type (week, month, lifetime) */
  plan: string;
  /**
   * ISO timestamp when the license expires (null for lifetime)
   * @nullable
   */
  expiresAt: string | null;
}

export interface LicenseStatusResponse {
  valid: boolean;
  plan: string;
  /** @nullable */
  expiresAt: string | null;
}

export interface CreateLicenseBody {
  /** Plan type - week, month, or lifetime */
  plan: string;
  /** Admin secret key */
  adminSecret: string;
}

export interface LicenseCode {
  id: number;
  code: string;
  plan: string;
  /** @nullable */
  expiresAt: string | null;
  activated: boolean;
  createdAt: string;
}

export interface RobloxAuthBody {
  /** The .ROBLOSECURITY cookie value */
  cookie: string;
}

export interface RobloxProfile {
  id: number;
  name: string;
  displayName: string;
  description: string;
  /** @nullable */
  avatarUrl: string | null;
}

export interface RobloxGroup {
  id: number;
  name: string;
  description: string;
  memberCount: number;
  /** @nullable */
  thumbnailUrl: string | null;
  publicEntryAllowed: boolean;
  isLocked: boolean;
}

export interface RobloxGroupsResponse {
  groups: RobloxGroup[];
}

export interface RobloxGroupStats {
  id: number;
  name: string;
  memberCount: number;
  funds: number;
  joinPolicy: string;
  isLocked: boolean;
  publicEntryAllowed: boolean;
  description: string;
  /** @nullable */
  thumbnailUrl: string | null;
}

export interface GenerateClothingBody {
  /** Text description of the clothing design to generate */
  prompt: string;
  /** Image size - "1024x1024", "512x512", or "256x256" */
  size?: string;
}

export interface GenerateClothingResponse {
  /** Base64-encoded image data */
  b64_json: string;
  /** The prompt used for generation */
  prompt: string;
}

export interface FeaturedGroupItem {
  /** The Roblox group ID */
  groupId: number;
  /** Group name */
  name: string;
  /** Number of group members */
  memberCount: number;
  /** URL of the group icon */
  thumbnailUrl?: string | null;
  /** When the group was last active on the platform */
  lastActiveAt: string;
}

export interface FeaturedGroupsResponse {
  groups: FeaturedGroupItem[];
}

export interface UploadClothingBody {
  /** The Roblox group ID to upload to */
  groupId: number;
  /** Name for the clothing item */
  name: string;
  /** Description for the clothing item */
  description?: string;
  /** Base64-encoded image data */
  imageData: string;
  /** Type of clothing - "Shirt" or "Pants" */
  clothingType: string;
}

export interface UploadClothingResponse {
  /** The Roblox asset ID of the uploaded clothing */
  assetId: number;
  name: string;
  status: string;
}