# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Limited.Ink API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clothing/generate:
    post:
      operationId: generateClothing
      tags: [clothing]
      summary: Generate clothing image with AI
      description: Generates a clothing image using AI based on the provided text prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateClothingBody"
      responses:
        "200":
          description: Generated image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateClothingResponse"
        "400":
          description: Invalid prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clothing/upload:
    post:
      operationId: uploadClothing
      tags: [clothing]
      summary: Upload clothing to Roblox
      description: Uploads a generated clothing item to a Roblox group. Uses the server-side session cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadClothingBody"
      responses:
        "200":
          description: Upload successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadClothingResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    GenerateClothingBody:
      type: object
      properties:
        prompt:
          type: string
          description: Text description of the clothing design to generate
        size:
          type: string
          description: Image size - "1024x1024", "512x512", or "256x256"
      required:
        - prompt
    GenerateClothingResponse:
      type: object
      properties:
        b64_json:
          type: string
          description: Base64-encoded image data
        prompt:
          type: string
          description: The prompt used for generation
      required:
        - b64_json
        - prompt
...
    UploadClothingBody:
      type: object
      properties:
        groupId:
          type: integer
          description: The Roblox group ID to upload to
        name:
          type: string
          description: Name for the clothing item
        description:
          type: string
          description: Description for the clothing item
        imageData:
          type: string
          description: Base64-encoded image data
        clothingType:
          type: string
          description: Type of clothing - "Shirt" or "Pants"
      required:
        - groupId
        - name
        - imageData
        - clothingType
    UploadClothingResponse:
      type: object
      properties:
        assetId:
          type: integer
          description: The Roblox asset ID of the uploaded clothing
        name:
          type: string
        status:
          type: string
      required:
        - assetId
        - name
        - status