# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Limited.Ink API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Validates a license code against the server, binds it to the device fingerprint, and returns a signed token
 * @summary Verify a license code
 */
export const VerifyLicenseBody = zod.object({
  code: zod.string().describe("The license activation code"),
  deviceFingerprint: zod
    .string()
    .describe("A hash uniquely identifying the user's device\/browser"),
});

export const VerifyLicenseResponse = zod.object({
  token: zod.string().describe("Signed JWT session token"),
  plan: zod.string().describe("The plan type (week, month, lifetime)"),
  expiresAt: zod
    .string()
    .nullable()
    .describe("ISO timestamp when the license expires (null for lifetime)"),
});

/**
 * Validates the device-bound JWT token using Authorization and X-Device-Fingerprint headers
 * @summary Validate an existing token via headers
 */
export const GetLicenseStatusHeader = zod.object({
  Authorization: zod.string().describe("Bearer <token>"),
  "X-Device-Fingerprint": zod
    .string()
    .describe("Raw device fingerprint computed client-side"),
});

export const GetLicenseStatusResponse = zod.object({
  valid: zod.boolean(),
  plan: zod.string(),
  expiresAt: zod.string().nullable(),
});

/**
 * Validates the device-bound JWT token and returns its status
 * @summary Validate an existing session token
 */
export const CheckLicenseStatusBody = zod.object({
  token: zod.string().describe("The JWT session token to validate"),
  deviceFingerprint: zod
    .string()
    .describe("The device fingerprint to match against the token"),
});

export const CheckLicenseStatusResponse = zod.object({
  valid: zod.boolean(),
  plan: zod.string(),
  expiresAt: zod.string().nullable(),
});

/**
 * Validates the .ROBLOSECURITY cookie, stores it in the server session, and returns the user profile. The cookie is never returned to the client after this call.
 * @summary Authenticate with Roblox cookie
 */
export const RobloxAuthBody = zod.object({
  cookie: zod.string().describe("The .ROBLOSECURITY cookie value"),
});

export const RobloxAuthResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  displayName: zod.string(),
  description: zod.string(),
  avatarUrl: zod.string().nullable(),
});

/**
 * Destroys the server-side Roblox session, clearing the stored cookie
 * @summary Logout Roblox session
 */
export const RobloxSessionLogoutResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns recently active Roblox groups on the platform. Public endpoint, no authentication required.
 * @summary Get featured groups
 */
export const GetFeaturedGroupsResponse = zod.object({
  groups: zod.array(
    zod.object({
      groupId: zod.number().describe("The Roblox group ID"),
      name: zod.string().describe("Group name"),
      memberCount: zod.number().describe("Number of group members"),
      thumbnailUrl: zod.string().nullish().describe("URL of the group icon"),
      lastActiveAt: zod
        .date()
        .describe("When the group was last active on the platform"),
    }),
  ),
});

/**
 * Returns Roblox groups where the authenticated user is an Owner. Uses the server-side session cookie.
 * @summary Get owner groups
 */
export const GetRobloxGroupsResponse = zod.object({
  groups: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      description: zod.string(),
      memberCount: zod.number(),
      thumbnailUrl: zod.string().nullable(),
      publicEntryAllowed: zod.boolean(),
      isLocked: zod.boolean(),
    }),
  ),
});

/**
 * Returns detailed statistics for a specific Roblox group. Uses the server-side session cookie.
 * @summary Get group statistics
 */
export const GetRobloxGroupStatsParams = zod.object({
  groupId: zod.coerce.string(),
});

export const GetRobloxGroupStatsResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  memberCount: zod.number(),
  funds: zod.number(),
  joinPolicy: zod.string(),
  isLocked: zod.boolean(),
  publicEntryAllowed: zod.boolean(),
  description: zod.string(),
  thumbnailUrl: zod.string().nullable(),
});

/**
 * Generates a clothing image using AI based on the provided text prompt
 * @summary Generate clothing image with AI
 */
export const GenerateClothingBody = zod.object({
  prompt: zod
    .string()
    .describe("Text description of the clothing design to generate"),
  size: zod
    .string()
    .optional()
    .describe('Image size - \"1024x1024\", \"512x512\", or \"256x256\"'),
});

export const GenerateClothingResponse = zod.object({
  b64_json: zod.string().describe("Base64-encoded image data"),
  prompt: zod.string().describe("The prompt used for generation"),
});

/**
 * Uploads a generated clothing item to a Roblox group. Uses the server-side session cookie.
 * @summary Upload clothing to Roblox
 */
export const UploadClothingBody = zod.object({
  groupId: zod.number().describe("The Roblox group ID to upload to"),
  name: zod.string().describe("Name for the clothing item"),
  description: zod
    .string()
    .optional()
    .describe("Description for the clothing item"),
  imageData: zod.string().describe("Base64-encoded image data"),
  clothingType: zod
    .string()
    .describe('Type of clothing - \"Shirt\" or \"Pants\"'),
});

export const UploadClothingResponse = zod.object({
  assetId: zod
    .number()
    .describe("The Roblox asset ID of the uploaded clothing"),
  name: zod.string(),
  status: zod.string(),
});

/**
 * Creates a new license code for a given plan type
 * @summary Create a new license code (admin)
 */
export const CreateLicenseBody = zod.object({
  plan: zod.string().describe("Plan type - week, month, or lifetime"),
  adminSecret: zod.string().describe("Admin secret key"),
});