# lib/api-zod/src/generated/types/uploadClothingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Limited.Ink API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UploadClothingBody {
  /** The Roblox group ID to upload to */
  groupId: number;
  /** Name for the clothing item */
  name: string;
  /** Description for the clothing item */
  description?: string;
  /** Base64-encoded image data */
  imageData: string;
  /** Type of clothing - "Shirt" or "Pants" */
  clothingType: string;
}